Extract QuestHelper from QuestHelperPlugin
This manager handles starting up and closing down helpers.
